+++ lib/Reader.php
@@ -36,12 +36,13 @@ class Reader {
      *
      * @param string|resource $data
      * @param int $options
-     *
+     * @param string $charset
      * @return Document
      */
-    static function read($data, $options = 0) {
+    static function read($data, $options = 0, $charset = 'UTF-8') {
 
         $parser = new Parser\MimeDir();
+        $parser->setCharset($charset);
         $result = $parser->parse($data, $options);
 
         return $result;